Separação do interpretador do interpretador base.
