+++ fontes/interpretador/index.ts
@@ -1,4 +1,5 @@
 export * from './erro-interpretador';
 export * from './interpretador';
+export * from './interpretador-base';
 export * from './interpretador-com-depuracao';
 export * from '../interfaces/retornos/retorno-interpretador';
